+++ common/src/main/java/dev/mayaqq/estrogen/datagen/recipes/EstrogenRecipeInterface.java
@@ -2,6 +2,8 @@ package dev.mayaqq.estrogen.datagen.recipes;
 
 import com.google.gson.JsonArray;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.item.Item;
 
 public interface EstrogenRecipeInterface {
     long getAmount(long amount);
@@ -13,4 +15,6 @@ public interface EstrogenRecipeInterface {
     EstrogenLoadCondition isModLoaded(String modid);
 
     public record EstrogenLoadCondition(String name, JsonArray condition) {}
+
+    public TagKey<Item> getCommonTag(String name);
 }