+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Estrogen `@VERSION@` `1.20.1` `Fabric & Forge`
 ### Changelog:
-- Chest second layer
-- Crash and bug fixes
+- Fixed issue with infinite dashes
+- Fixed Patches being too good
+- Cookie Jar Recipe